Refactored/Fixed/Added inputing by CW paddles & USB keyboard & F-butt…

+++ mchf-eclipse/drivers/audio/cw/cw_gen.h
@@ -17,16 +17,17 @@
 #include "arm_math.h"
 
 // Exports
-void 	CwGen_Init();
+void    CwGen_Init();
+
 void    CwGen_PrepareTx();
 void    CwGen_SetSpeed();
 
 bool    CwGen_TimersActive();
-bool	CwGen_Process(float32_t *i_buffer,float32_t *q_buffer,ulong size);
+bool    CwGen_Process( float32_t *i_buffer, float32_t *q_buffer, uint32_t size );
 
-void 	CwGen_DahIRQ();
-void 	CwGen_DitIRQ();
+void    CwGen_DahIRQ();
+void    CwGen_DitIRQ();
 
-uint8_t CwGen_CharacterIdFunc(uint32_t);
+uint8_t CwGen_CharacterIdFunc( uint32_t );
 
 #endif
